Feat : "기본 기능 완성	"

+++ subject1/subject/src/main/java/com/subject/subject/service/ReviewService.java
@@ -19,7 +19,7 @@ public class ReviewService {
                 .build();
     }
 
-    public RespReviewDto getReviews(int movieId) {
+    public RespReviewDto getReviews(Long movieId) {
         return reviewMapper.getReviewsByMovieId(movieId);
 
     }
